requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
requests==2.32.5
python-dotenv==1.2.1
pydantic==2.12.5

# LLMs & AI
groq==0.37.1
torch==2.10.0

# LangChain Ecosystem
langchain==1.2.10
langchain-core==1.2.14
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-text-splitters==1.1.1
langchain-groq==1.1.2
langchain-huggingface==1.2.0

# LangGraph
langgraph==1.0.9
langgraph-cli==0.4.9

# Vector Database
pinecone==7.3.0
langchain-pinecone==0.2.13

# Database & Document Processing
langgraph-checkpoint-postgres==3.0.4
psycopg-pool==3.3.0
psycopg-binary==3.3.3
pypdf==6.7.1